+++ index.html
@@ -23,6 +23,7 @@
 
 		<script src="./script/logic/LocalStorage/addNewTask.js" type="module"></script>
 		<script src="./script/logic/LocalStorage/deleteTask.js" type="module"></script>
+		<script src="./script/logic/LocalStorage/addTaskButton.js" type="module"></script>
 		<script src="./script/main.js" type="module"></script>
 	</body>
 </html>
