+++ script/logic/LocalStorage/naiming.js
@@ -2,3 +2,10 @@ export const nameTasks = "tasks";
 export const nameTextTasks = "text";
 export const nameChecedTasks = "checked";
 export const nameidTasks = "id";
+
+const clearInputButton = document.getElementById("clear-input-button");
+const newTaskInput = document.getElementById("new-task-input");
+const taskList = document.getElementById("taskList");
+const addTaskButton = document.getElementById("add-task-button");
+
+export { clearInputButton, newTaskInput, taskList, addTaskButton };
